requirements.txt
requests
beautifulsoup4
functions-framework 
google-cloud-storage
sendgrid

